Add section anchors
This will probably mess with the CSS, but I find myself wanting to link to sections directly (especially for larger pagers).  If you link this as is, cool, or want to check on the CSS, either way.  I can bang out a few more anchors once you give the general OK.