[installation] set weight of 8 for upgrading doc
When I brought over the upgrading doc from the legacy content, I
introduced some problems. The existing doc lacked a `weight`
specification in the document metadata, leading to some errors in
publishing, and prompting us to set a default weight of 0 in
https://github.com/sensu/sensu-docs/pull/554/.

I think this default weight needs to be revisited, but as a first step
lets set an appropriate weight for the upgrade docs.