update the doc to be more clear that redact config is override and not merge.
This was brought up in slack that it was not obvious to them that this was an override and assumed it was a merge operation. Regardless of changing this behavior if we want to (I don't think we should) I think we should minimally make the documention a bit clearer on this. We might also want to look at client at startup and emit a warning but that is out of the scope of this.

Signed-off-by: Ben Abrams <me@benabrams.it>